# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/feedback-requests": {
      "post": {
        "operationId": "FeedbackRequestsController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackRequestsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackRequestsEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Feedback Requests"
        ]
      },
      "get": {
        "operationId": "FeedbackRequestsController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FeedbackRequestsEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Feedback Requests"
        ]
      }
    },
    "/api/v1/feedback-requests/{id}": {
      "get": {
        "operationId": "FeedbackRequestsController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackRequestsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Feedback Requests"
        ]
      },
      "patch": {
        "operationId": "FeedbackRequestsController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequestsEntity"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackRequestsEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Feedback Requests"
        ]
      }
    },
    "/api/v1/main-config": {
      "post": {
        "operationId": "MainConfigController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMainConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainConfigDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Main Config"
        ]
      },
      "get": {
        "operationId": "MainConfigController_findMany_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainConfigDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Main Config"
        ]
      }
    },
    "/api/v1/main-config/{id}": {
      "get": {
        "operationId": "MainConfigController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainConfigDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Main Config"
        ]
      },
      "patch": {
        "operationId": "MainConfigController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMainConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MainConfigDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Main Config"
        ]
      }
    },
    "/api/v1/main-config/key/{config_key}": {
      "get": {
        "operationId": "MainConfigController_findByKey_api/v1",
        "parameters": [
          {
            "name": "config_key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MainConfigDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Main Config"
        ]
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "operationId": "OrdersController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrdersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrdersDto"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "get": {
        "operationId": "OrdersController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrdersEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/pre-orders": {
      "post": {
        "operationId": "PreOrdersController_create_api/v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePreOrdersDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreOrdersEntity"
                }
              }
            }
          }
        },
        "tags": [
          "Pre Orders"
        ]
      },
      "get": {
        "operationId": "PreOrdersController_findAll_api/v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreOrdersEntity"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Pre Orders"
        ]
      }
    },
    "/api/v1/pre-orders/{id}": {
      "get": {
        "operationId": "PreOrdersController_findOne_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreOrdersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Pre Orders"
        ]
      },
      "patch": {
        "operationId": "PreOrdersController_update_api/v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePreOrdersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreOrdersEntity"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Pre Orders"
        ]
      }
    }
  },
  "info": {
    "title": "Custom Admin Panel API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateFeedbackRequestsDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "hide_content": {
            "type": "string",
            "nullable": true
          },
          "user_full_name": {
            "type": "string",
            "nullable": true
          },
          "user_email": {
            "type": "string",
            "nullable": true
          },
          "user_address": {
            "type": "string",
            "nullable": true
          },
          "user_phone": {
            "type": "string",
            "nullable": true
          },
          "from_path": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "FeedbackRequestsEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          },
          "hide_content": {
            "type": "string",
            "nullable": true
          },
          "user_full_name": {
            "type": "string",
            "nullable": true
          },
          "user_email": {
            "type": "string",
            "nullable": true
          },
          "user_address": {
            "type": "string",
            "nullable": true
          },
          "user_phone": {
            "type": "string",
            "nullable": true
          },
          "from_path": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "update_at": {
            "type": "string",
            "format": "date-time"
          },
          "processed": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "title",
          "content",
          "hide_content",
          "user_full_name",
          "user_email",
          "user_address",
          "user_phone",
          "from_path",
          "created_at",
          "update_at",
          "processed"
        ]
      },
      "CreateMainConfigDto": {
        "type": "object",
        "properties": {
          "config_key": {
            "type": "string",
            "description": "Key of the configuration setting",
            "enum": [
              "currency_code",
              "currency_symbol",
              "product_weight_primary_unit",
              "product_weight_secondary_unit",
              "product_dimensions_unit",
              "shipping_calculating_service",
              "payment_service",
              "business_name",
              "business_company",
              "business_street1",
              "business_street2",
              "business_city",
              "business_state",
              "business_zip",
              "business_country",
              "business_phone",
              "business_email"
            ],
            "example": "currency_code"
          },
          "config_value": {
            "type": "string",
            "description": "Value of the configuration setting",
            "enum": [
              "usd",
              "$",
              "lb",
              "oz",
              "in",
              "shippo",
              "stripe"
            ],
            "example": "usd"
          }
        },
        "required": [
          "config_key",
          "config_value"
        ]
      },
      "MainConfigDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "config_key": {
            "type": "string",
            "description": "Key of the configuration setting",
            "enum": [
              "currency_code",
              "currency_symbol",
              "product_weight_primary_unit",
              "product_weight_secondary_unit",
              "product_dimensions_unit",
              "shipping_calculating_service",
              "payment_service",
              "business_name",
              "business_company",
              "business_street1",
              "business_street2",
              "business_city",
              "business_state",
              "business_zip",
              "business_country",
              "business_phone",
              "business_email"
            ],
            "example": "currency_code"
          },
          "config_value": {
            "type": "string",
            "description": "Value of the configuration setting",
            "enum": [
              "usd",
              "$",
              "lb",
              "oz",
              "in",
              "shippo",
              "stripe"
            ],
            "example": "usd"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "update_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "config_key",
          "config_value",
          "created_at",
          "update_at"
        ]
      },
      "UpdateMainConfigDto": {
        "type": "object",
        "properties": {
          "config_key": {
            "type": "string",
            "description": "Key of the configuration setting",
            "enum": [
              "currency_code",
              "currency_symbol",
              "product_weight_primary_unit",
              "product_weight_secondary_unit",
              "product_dimensions_unit",
              "shipping_calculating_service",
              "payment_service",
              "business_name",
              "business_company",
              "business_street1",
              "business_street2",
              "business_city",
              "business_state",
              "business_zip",
              "business_country",
              "business_phone",
              "business_email"
            ],
            "example": "currency_code"
          },
          "config_value": {
            "type": "string",
            "description": "Value of the configuration setting",
            "enum": [
              "usd",
              "$",
              "lb",
              "oz",
              "in",
              "shippo",
              "stripe"
            ],
            "example": "usd"
          }
        },
        "required": [
          "config_key",
          "config_value"
        ]
      },
...
      "CreateOrdersDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string"
          },
          "payment_id": {
            "type": "string"
          },
          "order_status": {
            "type": "string"
          },
          "order_amount_subtotal": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_total": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_discount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_shipping": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_tax": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_items": {
            "type": "object"
          },
          "payment_status": {
            "type": "string"
          },
          "shipping_order_id": {
            "type": "string",
            "nullable": true
          },
          "shipping_details": {
            "type": "object"
          },
          "shipping_options": {
            "type": "object"
          }
        },
        "required": [
          "user_id",
          "payment_id",
          "order_status",
          "order_items",
          "payment_status",
          "shipping_details",
          "shipping_options"
        ]
      },
...
      "OrdersEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "users": {
            "$ref": "#/components/schemas/UsersEntity"
          },
          "payment_id": {
            "type": "string"
          },
          "order_number": {
            "type": "integer",
            "format": "int32"
          },
          "order_status": {
            "type": "string"
          },
          "order_amount_subtotal": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_total": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_discount": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_shipping": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_amount_tax": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "order_items": {
            "type": "object"
          },
          "payment_status": {
            "type": "string"
          },
          "shipping_order_id": {
            "type": "string",
            "nullable": true
          },
          "shipping_details": {
            "type": "object"
          },
          "shipping_options": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "update_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "user_id",
          "payment_id",
          "order_number",
          "order_status",
          "order_amount_subtotal",
          "order_amount_total",
          "order_amount_discount",
          "order_amount_shipping",
          "order_amount_tax",
          "order_items",
          "payment_status",
          "shipping_order_id",
          "shipping_details",
          "shipping_options",
          "created_at",
          "update_at"
        ]
      },
...
      "CreatePreOrdersDto": {
        "type": "object",
        "properties": {
          "product": {
            "type": "object"
          },
          "product_quantity": {
            "type": "integer",
            "format": "int32"
          },
          "user_address": {
            "type": "string"
          },
          "user_full_name": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "product",
          "product_quantity",
          "user_address",
          "user_full_name",
          "user_email"
        ]
      },
      "PreOrdersEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "product": {
            "type": "object"
          },
          "product_quantity": {
            "type": "integer",
            "format": "int32"
          },
          "user_address": {
            "type": "string"
          },
          "user_full_name": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "product",
          "product_quantity",
          "user_address",
          "user_full_name",
          "user_email",
          "content"
        ]
      },
      "UpdatePreOrdersDto": {
        "type": "object",
        "properties": {
          "product": {
            "type": "object"
          },
          "product_quantity": {
            "type": "integer",
            "format": "int32"
          },
          "user_address": {
            "type": "string"
          },
          "user_full_name": {
            "type": "string"
          },
          "user_email": {
            "type": "string"
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        }
      }
    }
  }
}